chore: modify test to work with new PrepareProposal
